# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Tone of the content (professional, friendly, persuasive)
 */
export type SeoInputTone = typeof SeoInputTone[keyof typeof SeoInputTone];


export const SeoInputTone = {
  professional: 'professional',
  friendly: 'friendly',
  persuasive: 'persuasive',
} as const;

export interface SeoInput {
  /**
     * Name of the service or product to generate SEO for
     * @minLength 1
     * @maxLength 200
     */
  serviceName: string;
  /** Language for generated content (e.g. "ar", "en") */
  language?: string;
  /** Tone of the content (professional, friendly, persuasive) */
  tone?: SeoInputTone;
}

export interface FaqItem {
  question: string;
  answer: string;
}

export interface SeoResult {
  /** SEO page title (50-60 chars) */
  title: string;
  /** Meta description (150-160 chars) */
  metaDescription: string;
  /** List of SEO keywords */
  keywords: string[];
  /** Short catchy slogan for the service */
  slogan: string;
  /** Open Graph title */
  ogTitle: string;
  /** Open Graph description */
  ogDescription: string;
  /** Twitter card title */
  twitterTitle: string;
  /** Twitter card description */
  twitterDescription: string;
  /** URL-friendly slug for canonical URL */
  canonicalSlug: string;
  /** Frequently asked questions for schema markup */
  faqItems: FaqItem[];
}

export interface ErrorResponse {
  error: string;
}