# lib/api-zod/src/generated/types/seoInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoInputTone } from './seoInputTone';

export interface SeoInput {
  /**
     * Name of the service or product to generate SEO for
     * @minLength 1
     * @maxLength 200
     */
  serviceName: string;
  /** Language for generated content (e.g. "ar", "en") */
  language?: string;
  /** Tone of the content (professional, friendly, persuasive) */
  tone?: SeoInputTone;
}