# lib/api-zod/src/generated/types/seoResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FaqItem } from './faqItem';

export interface SeoResult {
  /** SEO page title (50-60 chars) */
  title: string;
  /** Meta description (150-160 chars) */
  metaDescription: string;
  /** List of SEO keywords */
  keywords: string[];
  /** Short catchy slogan for the service */
  slogan: string;
  /** Open Graph title */
  ogTitle: string;
  /** Open Graph description */
  ogDescription: string;
  /** Twitter card title */
  twitterTitle: string;
  /** Twitter card description */
  twitterDescription: string;
  /** URL-friendly slug for canonical URL */
  canonicalSlug: string;
  /** Frequently asked questions for schema markup */
  faqItems: FaqItem[];
}